% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRanalysis.R
\name{Get_GenInverse}
\alias{Get_GenInverse}
\title{The inverse of the matrix}
\usage{
Get_GenInverse(V.item)
}
\arguments{
\item{V.item}{The matrix.}
}
\value{
Output the inverse of the matrix.
}
\description{
Calculate the inverse of the matrix.
}
