% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.list.R
\docType{data}
\name{gwas.list}
\alias{gwas.list}
\title{GWAS files}
\format{
A list of 10 sex-stratified GWAS files. Each component of the list refers to one GWAS file. Each GWA study file has mandatory column headers:
\describe{
\item{MARKERNAME}{snp name}
\item{EA}{effect allele}
\item{NEA}{non effect allele}
\item{EAF}{effect allele frequency}
\item{CHROMOSOME}{chromosome of marker}
\item{POSITION}{position of marker}
\item{BETA}{allelic effect}
\item{SE}{standard error}
}
}
\usage{
gwas.list
}
\description{
The genome-wide association studies (GWAS) files of the 10 sex-stratified continental Africa cohorts.
}
\keyword{datasets}
