% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld.list.R
\docType{data}
\name{ld.list}
\alias{ld.list}
\title{The linkage disequilibrium (LD) matrices}
\format{
A list containing 5 LD correlation matrices. Each component of the list refers to one LD matrix.
The name of each component refers to one Afican cohort. The five cohorts are ESN, GWD, LWK, MSL and YRI.
}
\usage{
ld.list
}
\description{
The LD correlation matrices from 5 continental African cohorts without sex-stratification.
}
\keyword{datasets}
