% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envMRMEGA.R
\name{lr_w}
\alias{lr_w}
\title{The weighted linear regression.}
\usage{
lr_w(Y, X, W)
}
\arguments{
\item{Y}{BETA values of the genetic variants across all cohorts.}

\item{X}{PCs.}

\item{W}{The vector of weights, each component referring to the estimated inverse variance of the allele effect at the corresponding variants. The length of the vector is the number of cohorts.}
}
\value{
Output the data frame containing information of fitting the regression model: the estimated coefficients, standard errors, the deviance of the different constrained  models and the corresponding degrees of freedom.
}
\description{
Fit the linear regression with weights using GWAS data across cohorts.
The contribution of each GWAS is weighted by the estimated inverse variance of the reference allele effect at the corresponding variant.
}
\author{
Siru Wang
}
