% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRMEGAout.R
\docType{data}
\name{MRMEGAout}
\alias{MRMEGAout}
\title{The meta-analysis output from MR-MEGA}
\format{
A dataframe with 301 genetic variants and 17 variables:
\describe{
\item{MARKERNAME}{unique marker identification across input files.}
\item{beta0}{effect of the intercept of meta-regression.}
\item{se0}{std error of the intercept of meta-regression.}
\item{beta_i}{effect of the i-th PC of meta-regression.}
\item{se_i}{std error of the effect of the i-th PC of meta-regression.}
\item{chisq_association}{chisq value of the association.}
\item{ndf_association}{the number of degrees of freedom of the association.}
\item{pvalue_association}{p-value of the association.}
\item{chisq_anceheter}{chisq value of the heterogeneity due to different ancestry.}
\item{ndf_anceheter}{the number of degrees of freedom of ancestral heterogeneity.}
\item{pvalue_anceheter}{p-value of the ancestral heterogeneity.}
\item{chisq_residual}{chisq value of the residual heterogeneity.}
\item{ndf_residual}{the number of degrees of freedom of the residual heterogeneity.}
\item{pvalue_residual}{p-value of the residual heterogeneity.}
\item{logBF}{log of Bayesian Factors.}
}
}
\usage{
MRMEGAout
}
\description{
A meta-analysis results of genetic variants across all the genome. This approach uses genome-wide metrics of diversity between populations
to derive axes of genetic variation via multi-dimensional scaling. Allelic effects of a variant across GWAS, weighted by their corresponding standard errors,
can then be modelled in a linear regression framework, including the axes of genetic variation as covariates.
}
\keyword{datasets}
