% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sexsmk.R
\docType{data}
\name{sexsmk}
\alias{sexsmk}
\title{The smoking proportions across the 10 sex stratified populations.}
\format{
A data frame containing smoking proportions and sex indicators across 10 cohorts.

\describe{
\item{sex}{sex indicates female and male (female=1, male=0).}
\item{smoke}{smoking proportions ranging from 20\% to 90\%.}
}
}
\usage{
sexsmk
}
\description{
The smoking proportions across the 10 sex stratified populations.
}
\keyword{datasets}
